// web/lib/types/writing-style.ts
/**
 * Enum for Writing Styles used in Publishing Studio
 */
export enum WritingStyle {
    PROFESSIONAL = 'Professional',
    FRIENDLY = 'Friendly',
    HUMOROUS = 'Humorous',
    PERSUASIVE = 'Persuasive',
    INSPIRATIONAL = 'Inspirational',
    TRENDY = 'Trendy',
    STORYTELLING = 'Storytelling',
    CONCISE = 'Concise',
    SALE_HARD = 'Sale Hard',
    NUTRITION_EXPERT = 'Nutrition Expert',
    MARKETING_EXPERT = 'Marketing Expert'
}